+++ CHANGES.rst
@@ -14,6 +14,7 @@ Here you can see the full list of changes between each SQLAlchemy-Continuum rele
 - One-To-One relationship support for reflected version class relationships
 - Removed tx_context context manager. Transaction objects can now be created manually and user has direct access to the parameters of this object.
 - Removed tx_meta context manager. Transaction meta objects can now be created explicitly.
+- Fixed association reverting when the relationship uses uselist=False
 - Deprecated extensions
 
 0.10.3 (2014-02-27)